move login and signup forms into components, use them in pages and modals
