+++ frontend/src/components/Forms/Form.module.css
@@ -1,0 +1,19 @@
+.formControl {
+    border: 2px solid var(--color-grey);
+}
+
+.formControl:active {
+    border: 2px solid var(--color-blue);
+    box-shadow: none;
+    background-color: #212529;
+}
+
+.formControl:focus {
+    border: 2px solid var(--color-blue);
+    box-shadow: none;
+    background-color: #212529;
+}
+
+.formGroup {
+    min-height: 100px;
+}
